reorganize session code into new file
